comment, rename, add treasury unit test
