+++ src/treasury.hl
@@ -1,0 +1,24 @@
+spending treasury
+
+// TODO: change file to Escrow, implement the Escrow validator
+// Find the Escrow validator here:
+// https://github.com/IntersectMBO/plutus-apps/blob/main/plutus-use-cases/src/Plutus/Contracts/Escrow.hs
+
+struct Datum {
+    creator: PubKeyHash
+}
+
+enum Redeemer {
+    Cancel
+}
+
+func main(datum: Datum, redeemer: Redeemer, context: ScriptContext) -> Bool {
+    tx: Tx = context.tx;
+
+    redeemer.switch {
+        Cancel => {
+            // Check that the owner signed the transaction
+            tx.is_signed_by(datum.creator).trace("VS2: ")
+        }
+    }
+}
